Fixed up activation panel
