+++ frontend/src/layout/navigation-3000/sidepanel/panels/activation/activationLogic.test.ts
